Move CollectionConfig under Api namespace
